+++ app/order/api.py
@@ -4,7 +4,7 @@ from app.order.order_model import Order
 from app.menu.menu_model import Menu
 from app.auth.auth_model import User
 from app.auth.decorator import token_required
-from app.validate import validate_order
+from app.order.order_helper import validate_order
 from app.database.database import Database
 from app.database.order_db_queries import OrderDbQueries
 from app.database.menu_db_queries import MenuDbQueries
