+++ app/order/order_helper.py
@@ -1,0 +1,15 @@
+"""This module contain functions that handle input validation"""
+import re
+def validate_order(data):
+    """validates the order inputs and return appropriate message"""
+    try:
+        if not data['item_name'].strip()  or not data['quantity'].strip():
+            return "all fields are required"
+        if not re.match("^[a-zA-Z]*$", data['item_name'].strip()):
+            return "order should only contain alphabetic characters "
+        elif not re.match("^[0-9_]*$", data['quantity'].strip()):
+            return "Quantity should be a number"
+        else:
+            return 'valid'
+    except KeyError:
+        return "All keys are required"
